. use d_mem for linux ,windws, ucos,  these os  do not have mem heap function . add should_stop flag for windows

+++ OSAL/free_rtos/osal.c
@@ -20,7 +20,7 @@ void os_start(void)
   vTaskStartScheduler();
 }
 //-----------------------------------------------------------------------------------------------------------
-void *os_malloc(u32 size)
+void *os_malloc(u16 size)
 {
   return pvPortMalloc(size);
 }
